Change how we get outgoing email addresses. Work around recent change. (#7)
The work-around is for issue #6012, but it only works in
this branch until it gets pulled since it uses the new
setting/feature.
Also change how we get email addresses so we can get every email
address and reply to that the user has, instead of only the
primary one. This also sets the reply to correctly.