Feature/allow users to email using system account as themselves (#13)
With this option enabled, users will be able to send email out
of SuiteCRM with their name and email address, but sending though
the system outgoing email server. Their email address and name
will be populated in the "From" dropdown.
Clean up the text, and shorten the LBL_ALLOW_DEFAULT_SELECTION
label so it doesn't mess with the formatting.

Also change how we get email addresses so we can get every email
address and reply to that the user has, instead of only the
primary one. This also sets the reply to correctly.